+++ src/transporters/nats.js
@@ -196,7 +196,6 @@ class NatsTransporter extends Transporter {
 		if (!this.client) return Promise.resolve();
 
 		return new Promise(resolve => {
-			this.incStatSent(data.length);
 			this.client.publish(topic, data, resolve);
 		});
 	}